Re-launched the bunch of validation tests.
Now, the coupling nse>cd is clean after having managed the interpolation thing.
There are still some minor differences in the tests 32 and 33, between coupling cd>nse activated and deactivated. Needs further investigation to be sure that the code is correct: do like it was done in the previous commits 1976 to 1978: display analytical values and Param values of the viscosity and check it.

This is the next step. + make the interpolation correctly as it was done for Time_cd2d.

Now another point that was tested in this commit but then erased for lisibility purpose: when we activate AFC, a parameter FULL_INTERNAL_MATRIX is put to 1, and it destroys the BlockFEMatrix of TNSE2D, even when TCD2D is not solved! this is because of global dependency...

This issue is the main remaining issue to get AFC work correctly in the program.
But first, the above step