Oh and I forgot in the last commit to say that the examples 30 and 31 were re-launched and the values at the quadrature points are displayed, as it was tested in early commits with TCD2D.
Namely, the values of the analytical viscosity at the quadrature points were compared with those of the viscosity field coming from tcd2d.get_solution(). Result: it matches but  not perfectly and the difference is due to the interpolation.

So, the example "kind of" validates the coupling cd>nse so far, and keep in mind that the less regular the viscosity field is, the bigger the error of TNSE2D might be. If we use analytical viscosity which have very strong gradients, such as the ones implemented in Beltrami example 32, then, it is expected that the difference between interpolated viscosity and analytical values is quite high..this will be done in later commits, when we are done with the current interpolation issue (where we just focus on example 31 for simplicity of programming and debugging).