Failed attempt to make one interpolation for the whole system class Time_NSE2D...
In this commit, nothing happened, just some esthetic and warning removal.

Difficult to set one interpolation member in the class, because FEFunction needs initialization, which can be done only in System_Per_Grid. It means, definition of one interpolated rho and mu per system_per_grid in case we use multigrid....the whole rho and mu_field should be re-done then.

We stay with other temporary solution: make the interpolation each time we assemble...it is gonna be heavy, but the refactoring of the whole class will take some time, so not now.

Next step: finish interpolation work by adding in it the mass_matrix assemble routine. Then test (on cheap examples) the interpolation, on several time step, with 30, 31 and 32