Tested with success Beltrami 32 with different values of P1, P2, P3 as well as Ansatz_order=1,2,3.
A few commits earlier, changing the Ansatz_order to 1 or 3 would have failed. Now, thanks to the interpolation done on the previous commits, it works, although it is too ugly to make one interpolation at each assembling routine...this will need definitely a correction.

The comparison was done between all the booleans = false, and between all boolean true (except coupling nse>cd because it is not the purpose in these examples). There are quite some difference between analytical value and assembled values of the variable viscosity mu, especially when there is a strong gradient...

But the coupling and the interpolation seem to be quite reliable. The value differences come from numerical error of interpolation, discretization...

Next step: just check quickly the story with LaplaceType1, and adapt the rhs of the Beltrami example..