WORKAROUND DONE FOR AFC!
Now, AFC works fine with the coupling nse>cd! The only remaining problem to solve is the long interpolation time. If this is solved, then the code is very good.

It is not a nice workaround, I just changed the parameter INTERNAL_FULL_MATRIX_STRUCTURE back to 0 before each TNSE2D methods and back to 1 before each TCD2D.

Tested on the rotating semi-circle, for 3s simulation, refinement step 4, it works fine...but too many and too long interpolation...

Next step: go back to interpolation issue or go further with IMEX...