Finished implementation of IMEX
A quick test by puttin input parameter "time_discretization" to 4 with the coupling CD>NSE on example 31 was positive....the code needs certainly to be rewritten to improve its lisibility. So far it was just a copy-past of the IMEX in TNSE3D. The parameter "time_discretization" is not good managed yet, it just reacts when it is 4.

It was found that there is no flow around cyliner benchmark example for TNSE2D. SO THE IMEX IS NOT OFFICIALLY VALIDATED.

But because it was all copy-pasted, we assume it is ok.

At this state, we have a program where the coupling nse>cd was proven to work on 3 particular examples, where the coupling cd>nse was proven to work on also 3 other examples, where IMEX is implemented as in TNSE3D and not validated yet rigorosly with "flow around cylinder", where TCD2D can use AFC without disturbing TNSE2D, where different space orders between TNSE2D and TCD2D can be used thanks to the interpolation.

The main issues concern TNSE2D: it is just a mess, too much routines, with a bad management of local assembling objects with the fields rho, mu and u + interpolation everywhere and too costly interpolations.
TNSE2D MUST be optimized, at least to increase lisibility and re-usage of code.

On the contrary, TCD2D seems quite well organized and understandable, no special work on it is needed right now. What is missing is the routines in case of non-linear TCD2D, but that's another problem....which might be taken into account later on.


Next important issues are 5:

1) focus on the interpolation issue: rewrite them to reduce their calls, or find another cheaper way to do it.

2) think of how TNSE2D can be optimized (all the assemble routines with the rho and mu fields)

3) maybe run some test with 2-way-coupling with IMEX and AFC.

4) Find an analytical example with a viscosity variable in space AND TIME!

5) Find an analytical example for the 2-way coupling...

6) test benchmark problem, like Rayleigh-Taylor instability + dam break

7)..3D, parallel, multigrid...still a lot to do