In this commit, all the tests until dam break were re-done.
Examples 10 to 22 are successful.
Examples 30 and 31 are ok.
Example 32 Beltrami has shown discrepancies in the results. The bug was found and corrected: it came from the convection example.

However, other points have to be taken care of: for some combination of mu, mu_min and mu_max, the solution can explode! Two combined phenomena are suspected to cause this: when the coupling cd>nse is activated, convection equation is solved....and introduces SPURIOUS OSCILLATIONS! This could cause collapse of NSE. Second suspicion is that in some cases, the flow becomes non-viscous (high Reynolds number).

But the main bug was corrected, and now the viscosities mu1, mu2 and mu3 are correctly computed, constant in time and similar to a pure TCD2D computation.

Dam break: often, solution explodes. Probably comes from high Reynolds number. Also, another initial field was taken: a smooth gaussian, to see if the coupling was better, but still no-convergence. Finally, using AFC is working but interpolation is way too long.

To sum up the situation at this commit: it is urgent to debug dam break to obtain convergence, it is urgent to implement some stabilization in NSE for high reynolds number and it is urgent to solve the interpolation thing to make computation time reasonable.

Beltrami flow should be further used to test and debug, especially AFC with interpolation + convergence + error norms + NSE stabilization.

Next step: see with Volker about interpolation and NSE stabilization (Smago or VMS). Also, important to test at some time the IMEX with the flow around cylinder.