First commit in Raahe!
As described in earlier commits, there are several urgencies: study the convergence of the Beltrami problem (Navier-Stokes with variable viscosity), debug the dam break and debug interpolation problem.

Objective is to get a two-phase solver using AFC and IMEX, correct and operational as quick as possible.

Other things have to be done also, see earlier commits, not urgent now but should not be forgotten.

In this commit and the next ones, working on Beltrami flow: main program and input file are ready to launch tests.

Next step: check again the right hand side with Mathematica to be sure we evaluate the correct error norm.