Added Rayleigh-Taylor instability!
Example seems to work fine. This could validate the two-way coupling in the code.  When parameter P2 (rho_min) is equal to 1, viscous flow. Convergence OK.

In the code: mu_vector was forced to 1. Only density difference!

Bug to fix: in initial convection field, on the right side there is some irregularity appearing in Paraview...needs to be fixed.

Next step: make test with different rho_min, try to repeat images from Fraigneau et al, get rid of the oscillations due to convection equation.

This example should help to solve dam break convergence problem and further validate the implementation of the code.