MAJOR COMMIT!
The interpolation was worked around thx to a tip from Ulrich.

Now, no need for interpolations when different FERaeume are used.

The code was tested with one-way coupling on Rotating semi-circle (22) and then with two-way coupling on Rayleigh-Taylor. The output images seem to be correct. The current input file is ready to launch.

There are still major steps: focus on rayleigh-taylor to test convergence, then debug dam break convergence problem. Then the assembling methods are a mess, they need to be cleaned. Important code refactoring is expected. Then, focus on stabilization methods for TNSE (VMS?).

Next step: two-way coupling convergence check with rayleigh taylor and dam break.