Conclusion about Rayleigh-Taylor instability test.
Viscosity is constant = 1. Only density is variable, given an initial condition and then 2-way coupling.

The iterations converge (nonlinearmaxiteration was put to 20) and the images are good. A test case with rhomin=319 was solved on a coarse grid (level 5) to 3.7s, with dt=0.01. The images have been saved as a record.

The only bad news is that the images don't match with Fraigneau et al (2001), where they calculated the same problem with Re=1000 and had a kind of more viscous flow than the one here. I am quite sure, it has to do with the values of rhomin, domain size, right hand side of NSE, and other stuff like that...the code itself seems to be fine.

But it is probably not the current priority to make this example match the one by Fraigneau et al...maybe will discuss it later and check it another time. The main point here is that 2-way coupling works and iterations converge, when mu=1 and rho is variable and unknown.

Remember that in the Beltrami case, rho=1 and mu is variable (but given!).

Next step: focus on dam break and test convergence with 2-way coupling (try different cases: mu=1, rho=variable then mu= variable, rho=1, then both variable and unknown). Case 1 has been done for the pictures (in the previous commit I think) but there was only one iteration to make it quicker. Convergence has to be checked here.