Finished Beltrami analysis + reproduced RTInstability and dam break to check convergence.
The RTInstability converged, but the dam break not! In both cases, mu=1 and rho is variable. But in the first case, rho is not calculated with rho1 and rho2, it is just the solution of the convection equation.

We have here an interesting situation to debug dam break.

As it is, this commit allows to launch directly dam break and view its non-convergence.
In the next commit, I will just change enough to test dam break and check if there is convergence, and what are the differences..to be able to debug more easily.