MAJOR COMMIT. DAM BREAK FINALLY DEBUGGED.
In the example file, the parameter "INTERNAL PROJECT PRESSURE=1" should be 0...even when all BC are homogenous Dirichlet... then the problem converges...the reason for it is not known yet.

It is important to understand if this is correct and why it works. I just quickly checked and the parameter "INTERNAL PROJECT PRESSURE" doesnt seem to affect TCD2D, so it is purely a TNSE2D thing. How the variable density and viscosity in NSE require this parameter to be 0 when BC are homogeneous Dirichlet?..still to be explored.

Next step anyway: write the dam break cleaner, clean the code and do a real simulation to compare with the literature, using the rectangle geometry.