Tests successful.
After the merge and make check, the program has been compiled and the coupling tests were successful.

The code needs to be cleaned: management of the rho and mu variables depending on the example, input file difficult to read, automatisation of tests...etc

Next step: push to servers and continue working on Beltrami flow.