Beltrami debugged!
The coefficient "nu" in the coefficient was always 1 but should be equal to numin, so that the right hand side is 0 when numin=numax!!!!! The dimensions of the equations were not correct.

Next step: first focus on viscosity1, check a last time right hand side with Mathematica and then launch simulations in Optimode and check results for constant viscosities.