Corrected chrono objects in the program.
In this commit, a new bunch of simulations for Beltrami were done thanks to re-written and optimized Python scripts...

Tested with constant viscosity 0.1, 1 and 10. Note that after the debug in previous commits, the right hand side is now 0 and these 3 cases correspond exactly to a Beltrami flow. Now the curves should match the ones for Beltrami..but not sure because the L2U norm error is increasing with nu instead of decreasing...I will check with V.J.


Next step: continue Beltrami computations in Erhard servers  + focus on dam break benchmark...and oh yeah, push the code and compile in Opti in erhard + put there all the scripts stuff and prepare everything