Marked assembl2dSlipBC as old, and started new assembl2dslipBC with same input as assemble2d. (no discrete form, Taux...).
Commented out the SLIP condition in the standard Assemble2D.

Next step: correct the new assemble2dslipbc so that it works with new structure (e.g. with local assembling).

After this, the "old Assemble2dSlipBC" can be removed. And then, the assemble2dslipbc must be called in TNSE2D.C and tested.