TEMPORARY WORKAROUND FOR ASSEMBLE2DSLIPBC.
Because in standard cases, there is only one mass block matrix while the routine only takes into account 4 mass block matrices, the part of the routine where M12, M21 and M22 are modified have been commented in this commit. A quick test showed no segfault.

In the future, this has to be better coded, e.g. with the right number of blocks and matrix structure.

Next step: call the "apply_slip_bc" in the right places and check results on simple example.