NSType 4 for tnse2d with variable properties implemented.
Call to slip_bc, ok.

TODO: rewrite the variable properties of the fluid correctly (gather everything in local assembling + gather corresponding routines..)

TODO2: make a test file for all the examples and the different nstypes.

Next step anyway: make the call to apply_slip_bc in the correct places and test it with rayleigh-taylor