OK!
In this commit, the rand conditions were corrected. Problem was the number of rands. The geometry was so constructed that 0=bottom, 1 to x= right, x+1=top, x+2 to end= left.

Depending on the size of the rectangle box and the number of segments, the BCs had to be corrected accordingly.

Now, the geometry and BC are fine, identical to Chopet et al 2013, and free slip is active.
First simulations with coarse grid are not so bad...anyway better than early.

I think it is difficult to obtain same results, because the numerical methods are so different. So the example will stop here: just check again, for the last time, (and write it finally) the parameters of the Rayleigh taylor, the geometry and the BCs, launch a simulation (in OPTI mode) with fine grid, take nice pictures, and compare results (quickly) with Pochet et al, 2013 to validate the 2-phase flow with slip condition.