STOPED SLIP BC AND MOVE BACK TO LINEAR ELASTICITY.
In the last commit, the geometry and slip bc and everything was set up to reproduce Rayleigh Taylor as in Pochet et al, 2013. Simulations with (not so) fine grid was performed and output and vtks were saved for the report.

The next Baustelle for VOF_2Phase2D is implementation of a stabilization/turbulence model for TNSE2D. First = read the review from Naveed and Volker on VMS. Second= check implementation of VMS for TNSE2D. Three = implement it for the VOF_2Phase2D. Test will be done on dam break.

Now, linear elasticity: implement a velocity vector, initialize it, and construct the whole matrix system as needed for Newmark Scheme.