Added and initialize initial velocity.
Next step: think again on the necessity of implementing Newmark in this linear elasticity problem....is it what we need? how is it going to be used with the fluid structure interaction? should the model be used with a separate geometry for fluid and structure? decouple? should be the system posed as a saddle point system?...etc..