Starting implementation of VMS_Projection in Time_NSE2D.
So far, the VMS_Projection stuff is put only in the if-statement "!usingMultigrid". which means only the finest level is used. To implement the other grids, where SMAGORINSKY_COARSE is assembled, one should put the VMS_Projection and the systems_.emplace_back(..) BEFORE the if-statement "!usingMultigrid", this will construct the VMS_Projection and the finest level in any case, and then create the coarser grids, if multigrid is used.

Next step: continue implementing VMS_Projection in TNSE2D