Moved the "check" functions out from the ParameterDatabase class.
They are temporarily placed in Database.C, and are not member of any class. They mix global and local parameters, so it was found inappropriate to put them as members in the Database or ParameterDatabase classes.


They should be used when removing other global dependencies, like NSTYPE, LAPLACE_TYPE,...


Then, once they got rid of global parameters, they can be moved in a more relevant place.