VMS_PROJECTION IS ALMOST VALIDATED!
First comparison between "galerkin" and "vms_projection + TURBULENT_VICOSITY_CONSTANT=0" give similar result for example 0 (SinCos). Some small discrepancies are there, but in order of magnitude of e-15.


A second check with the more complex example 2 (SinCosExp, Beltrami 2D) revealed a big discrepancy in pressure between galerkin and vms with turbulence=0. This is not normal

Next step: debug the results and be sure to obtain the correct pressure and everything correct.