Reworked flow around cylinder to reproduce time-dependent benchmark problem.
Test of TNSE2D with Galerkin, dt=0.04, Q2/P1-disc, on reflevel2, with reynolds number 1000, as in Paper, reproduced same results (up to few decimals) as in paper.


The flow around cylinder is thus validated and can be used to test many other features, such as turbulence models, other time steppings, etc...

Next step in this branch is to test the PBVMS with the same input file and with TNSE2D to see if we obtain same results as galerkin, when turbulence=0.