Added VOF2D test file to be able to continue developing and changing the program in the near future.
Examples 10, 32 and 42 are tested, which correspond to different types of couplings.
Different geometry is also used, and the slip condition is also tested in example 42.

Example 22 needs to be debugged, there is a discrepancy and it looks it comes from the error computation. This needs further investigation.

Next step: clean the code, especially the assemble routines used in the couplings. this will make the code better and easier to move then to the 3D version on a good basis.

The turbulence with PBVMS has still to be tested on the flow around cylinder, with non zero turbulent constant.


Also consider implement Smagorinsky in 2D/3D just in case.