Rmoved assemble_non_linear_term_withfields.
Now, the variable rho and mu are used internally.


Also re-tested flow around cylinder with tnse2d and vms-projection with turbulence =0.0001. In the first steps it converges but after some steps, no convergence, stays at the same residual. Reynolds is still laminar 1000.


So the vms projection has not been proved to be functional yet...