Smagorinsky with FAC and high Re OK!
This is an empty commit, just to tell that Smagorinsky in TNSE2D worked fine with FAC with Re=100000 and turbulent visco_constant=0.01. The flow doesn't converge, is turbulent (diffused) and the drag, lift and deltaP coefficients have been plotted in folder Opti_tnse2d.


Next step: implement this turbulence model for VOF, namely the 2 routines (assemble initial time and assemble_nl_term) of smagorinsky. The rest should remain the same.