Here is a test of Dambreak with Smagorinsky!
Very coarse grid just to see if it runs ok. Turbulence was put to 0, but it stabilized already the flow, which is quite weird. But as usual, when the flow becomes complex, e.g. some vague or droplets want to leave the flow, then diverges, even when stabilized with Smago.


The remaining step to have it work properly is the surface tension.

Next step: try the bubble in the flow with Smagorinsky