Example of gas bubble in liquid tested with Opti_tmultiphase.
Launched with Smagorinsky, turbulence 0.01. After 8 time steps, the solution explodes.

The example file had to be adapted to obtain the bubble in unit squre instead of stirring in the ladle. In the input file, the position of the bubble was moved from bottom to middle to get a closed circle.

This was to illustrate that even with Smagorinsky, the code is not finished yet. Needs the surface tension...