Debugged 2 things: Needs2ndDerivatives was not set to "true" correctly.
Now it is ok. Second, the N_Parameters 4 was false, this must be 6.

Quick check shows that D11, D02 and D20 are 0, as expected. The interface in the example was put to horizontal line, so that the gradient must be only vertical. This is observed in D01. But a lot of values are not strictly 0, they are like e-15, in both vertical and horizontal directions, although D10 should be striclty 0 everywhere.


Further tests are needed to see effect of this.