Added FEFunction Phi to inject its gradient for the CSF model.
Added example 43 on equilibrium of rod to test the surface tension and compare with theoretical result deltaP = tau*kappa = tau/R.

Next step: test it and adjust coefficient of right hand side. This time, the CSF is calculated with grad.Phi and not grad.Rho. Phi is between 0 and 1, average on interface is supposed to be 0.5.

In paper Brackbill et al 1996, the radius is 2cm, which is too small for our unit squre. I took a bigger one, so the expected result cannot be compared with the paper. The geometry is not the same, but this should not change anything, since there is no gravity and the background fluid is at rest.