PROBLEM HERE WITH TRIANGLES.
First, the geo "TwoTriangles" work fine, no need for .GEO.
Second, even with triangles, the circle is bad, so the CSF will be bad, one needs unstructured grids.
Third, the whole NSE matrix was 0, the assembling was not made correctly with triangles. The problem probably comes from the multiplication by the phase field, which is P1...


The VOF needs debug with this triangles, P1-P2 etc...