THE CULPRIT HAS BEEN FOUND.
This commit is full of test, needs to be cleaned afterwards.


In triangles, d/dXidEta = 0 and in quads d/dXidEta!=0, so we obtain different second derivatives in the assembling routines between P1 and Q1, and thus, different curvatures and different results.


There is no error in the code. It is just incorrect to keep using a first-order phase field and calculate the CSF with it. One needs necessarily a second-order phase field...or a weak form for CSF, but that would not work with the volume integral model.


Next step: think of using interpolation to obtain a piece-wise C2 phase field out of the C1 used so far....or consider adapting the Algebraic Stabilisierung for second order fields ???