Smagorinksy with FAC in Multiphase program debugged!
Two things to correct: turbulent viscosity has to be scaled with rho, and rho should not be excessively high, e.g. instead of (rho,mu)=(100000,1), choose (1000,0.01). The first case converges much slower than the second case, and if the number of iterations is not big enough, the error in u can increase and the iterations could diverge after some time. This is avoided in the second case, because the  matrix system seems to be easier to be solved.


The convergence speed is a problem. One solution would be to divide the equation by rho to obtain a better conditioned matrix, but this would introduce two extra terms with grad(1/rho), which are non-zero in two-phase case. Will be the problem better then, with this implementation? numerical error? speed? Has to be discussed.



Next step: now that the Smagorinsky is ensured to be quite safe, re-test the dam break...CSF will be dealt with later, it is 0 right now.