ADDED LADLE GEOMETRY AND RE-WRITTEN THE EXAMPLE 50.
Now the geometry .mesh and .PRM works well for the ladle + the boundary functions are set up with appropriate boundary numbers. The climbing bubble can be reproduced also with this geometry.


Before this example, a whole bunch of tests for smagorisnky with the dam break were done. The turbulent VOF still needs to be improved.


Next step: correct the boundary functions (Param) to simulate gas injection in two areas within the bottom boundary. Without changing PRM and mesh.