Finished validation of code and test file.
The tests were reduced to one time step...the ideal would have been at least 2 time steps, but it was too long (>60s). The error tolerance was reduced from 1.e-7 to 1.e-6, because there is a small discrepancy between SEQ and MPI residuals.

Now that the main VOF program is there and is tested for uncoupled TNSE3D and TCD3D, next step = start implementation of coupling and changes needed for VOF. Maybe start with TCD3D because it is easier and quicker.