Added gas stirring in a cylindrical ladle 3D.
Still many problems : the geometry with the base circle is very bad, impossible to check the boundary conditions correctly + mumps and umfpack fail with refinement = 3 only, which is a very coarse grid + no fem-fct-cn in parallel yet.


So far, the plugs bc seem to be fine, added vtk output for initial solution, tried different combinations of parameters to obtain a nice and relevant geometry, but it is still very poor quality.

We should leave the example aside, and solve first the issue with algebraic stabilization in parallel. Then maybe focus on dam break example, or check what is wrong with the gas injection example.