Tested refinement level 4 in SEQ...
...using UMFPACK for TCD3D and FGMRES+LSC for TNSE3D. It is ca. 2x faster than UMFPACK for TNSE3D.

Reducing time step from 0.1 to 0.005 resulted to slight increase in computational time.

From this test campaign, it is concluded that it is impossible to use SEQ for a small 125kdof 3D problem, even with the best iterative solver.

It is necessary to have FEM-FCT-CN working in parallel, in order to launch tests in MPI, first with MUMPS, second using iterative solvers.


Next step:...plan and do the next step..