IMPORTANT COMMIT:Corrected input parameter for Iterative Solver!
Now, FGMRES+LSC is correctly set and it solves TNSE3D 10 times faster than Umfpack, and 5x times faster than in the previous commit, where the parameter "saddle_pp_velocity_solve" was put to "true".

Volker's input file was reproduced, setting this parameter on false, and the iterative solver is now very fast, which makes it possible to run the Refinement 4 (124kdof) on the UnitCube in SEQ on the laptop.

The program has been pushed, compiled and run now in Ehrard02, with Refinement 5, 940kdof, with this same solver, in SEQ.


Next step: take a look at resulting dam break, and start next big issue = FEM-FCT-CN in parallel...