Many problems with the gas injection example...
...the geometry can completely change the initial solution, even if the initial conditions are not changed...

The BCs seem now to be rather fine + the geometry found and used in this commit is also good.

Problem is that it first non linear iteration doesnt converge.

Next step: change BC of convection, e.g. remove gas injection and gas velocity, to see if we obtain the correct static pressure....if this is ok, then re-work injection BC in order to get it work...if needed, adapt the geometry...but this one is a good start anyway.