Some good news.
In the last commit, the pressure computed doesnt take into account the phases and densities...it is the pressure of a homogenous fluid in the whole domain, of density 1 and initial pressure is equal to static pressure = rho*g*h=1*-10*0.6 = 6.

In the commit before, the non-convergence is due to the initial velocity, which is too strong (or time/space grid too coarse). If the velocity =0, the iterations converge, and there is a slow convection due to density difference, the gas being induced by dirichlet condition in the bottom, without boundary velocity.

Next step: find a good time step and grid, to launch simulations of rising bubble with velocity.