Test runs now without column consistency update!
This behavior is still not clear: adding the Internal_full_matrix_structure feature in Assemble3D in a previous commit seems to result in a matrix with the correct column consistency!

The test runs now without this update, which can be seen in the duration of the test, 10s instead of >100s.

Mumps still needs to be debugged, and the results without column consistency update should be better understood...just be sure we are getting correct results.