Many failed tests for dam break, even with very fine time step 5e-7.
A lot of parameters were tried to get a convergence with a small viscosity ratio 1/0.5, but no way to get it work:
- damping (0.9) made a very tiny improvement in the iterations of the last steps...but then exploded. a smaller damping (0.7) made computation very unstable from the beginning.
- refined dt at given time steps, where the cfl warnings were observed to appear in preliminary tests....here also, iterations better converged at the end, but then exploded. dt was first refined to 5e-6, and even 5e-7 was tried. but in the last time steps where the solution exploded, the cfl warnung calculcated a dt=1e-12!
- max nl iterations = 50, gained a few time steps but then still diverges..
- different shapes of the dam to have a less sharp interface: curved corner and quarter of circle...but still a vague appears and diverges
- different size of the dam to create more space between top, and also on the right: when height smaller 40%, peak doesnt touch the top wall and is better, but still diverges at some moment
- different viscosites: 0.8 diverges later, other behavior
- different endtime: when a solution looked good at t=0.2, computation was run longer, in just one case, we ended up a stationary solution where velocity went to 0 and the liquid was 100% on the floor: the case where size and height = 0.03 and viscosity 0.8 and end time > 2s? in most cases, there was a moment where the solution diverges and the flow stopped suddenly with this weird peak.


Note that in almost cases, the standard mesh refinement was always 4. (finer = quicker divergence). The Boundary condition for convection equation is Neumann everywhere, while for NSE: slip condition on right, bottom and left + neumann on top (as seen in the example from OpenFOAM).


Conclusion of this tests campaign: it is probable that the VOF as implemented is not enough to model the dam break properly. Testing different meshes, turbulence, time steps, shapes...etc on the dam break didnt solve the problem. Maybe more work should be spent on the VOF, like e.g. implementing the CSF, and reconstructing the interface...and check the pressure condition because it can also be a reason. So far, the example runs without internal project pressure. but in some articles, the pressure was said to be important and one has to fix a value at some point of the mesh...


Hey, all this sounds very similar to the problem I had with comsol: no convergence while everything looked fine....maybe the pressure is the problem