Working commit for CSF.
Created function "compute_gradient" in FEFunction. This is a copy-paste from "interpolate" with the ability of storing the x and y components and it uses FindGradientLocal instead of FindGradient. Now, the gradient components of phi can be stored in FEFunctions.

Created function "smooth_gradient_phi" in Time_CD2D. This is a temporary function to re-construct a P1-version of gradient phi. Several fefunctions are created and compared (note that the function interpolate/compute_gradient can be used with P1 spaces instead of P0..in which case we obtain a P1 gradient phi without needing prolongation...but the values are a bit different than the prolongated gradient). Besides displaying the values and comparing visually the 0s, it is still difficult to check if the interpolated and prolongated functions are correctly computed.

Still a lot to do with this function: maybe re-write code using FEVectFunct, find a better way to compare results, add members to the objects to better manipulate the prolongated grad phi...etc.

This commit is far from being a final code version, but several things were done and it can be practical for the future to keep a trace of this work.

Note: two extra-changes: in the main program, call to smooth_gradient_phi() and in GridTransfer, commented out "delete aux" because it was causing fatal error when calling Prolongate.