Working commit 2 for CSF.
Introduced a TFEVectFunct with its corresponding BlockVector as members of Time_CD2D::System_Per_Grid in order to store the Gradient of Phi (assumed to be in same space as Phi, e.g. P1 after prolongation).

Saved the values of the gradient in the blockvector after using prolongate.
Checked the output by using WriteSol to both components of the FEVectFunct and comparing them with double* obtained after prolongation.

Now, the class Time_CD2D has the gradient of Phi in each system computed correctly in P1 (normally).

Next= check the derivative of the gradient, to check if it is non-zero + implement new members in Time_NSE2D where we transfer the FEVectFunct for further use in LocalAssembling objects and AssembleRoutines.