Quick check of last commit.
Computed gradient of the gradient, it is non-zero this time, so the results are considered to be fine so far.

Added a new member in Time_NSE2D to store the gradient of phi as a TFEVectFunct.

Next step: work on VOF_TwoPhase2D to save the gradient of phi into Time_NSE2D + work on Time_NSE2D and localassembling object to transfer the gradient of phi into assemble routines.