First interesting results of CSF!
Tested the drop pressure with this commit, using the new gradient of phi, one obtains a smoother circle.

Nothing was changed in the routine. It needs probably some adjustment (constants) and verification.

The only change was to put the gradient phi into local assembling object, change the LA appropriately, and that's it. The ParameterFct for rhs is unchanged and the routine for RHS also unchanged.

Next step: continue code verification