Failed test of Rayleigh Taylor Instability.
In this commit, put back the rayleigh taylor from Pochet et al + calculated the critical surface tension as given in Brackbill. Its value is 1.8.

Several surface tensions were tested: 0.5, 0.05, 0.005, but the problem diverges from first steps. In paraview, one can see waves appearing on the interface.


This is probably due to insufficiently fine mesh on the interface...one has to take care of the stability condition necessary for CSF. But since the mesh is already fine, no trial was done with a finer mesh, because too time-consuming...

Next step: try the original dam break, there, the mesh is quite fine, maybe it works.