Many failed tests and focus on Equilibrium rod benchmark.
Several configuration of dam break were tested, but the surface tension was unreallistically small and it introduced more iterations, and didnt make the flow much better.

With rayleigh taylor, not stable, probably needs finer interface.

Reproduced exact geometry of Rod as in Brackbill but with the given surface tension, solution explodes. Surface tension had to be turned down to 1e-8 to get a reasonable solution, although there are still waves and perturbations...which were not there in the Unit_Square test.


Tried different constants in the surface tension formula but nothing works. Impossible to get the right pressure in the rod.

Another difficulty: Brackbill uses grad of density while other paper uses grad of phi, and the constants are never the same....which one is correct?????

next step = solve the equilibrium rod to get correct results...then try rayleigh taylor again, and finally dam break.