Failed test of equilibrium rod.
Here, the correct surface tension was used, and the mesh and time step were refined very much until we obtain some convergence. Unfortunately, the circle is still waving and we are far from the theoretical pressure.

The coefficients in the surface tension were changed to make it even smaller, but no way to get it work.