FOUND ONE MISTAKE IN BRACKBILL ET AL!!!
In this commit, tried to obtain pressure 1.1805 as in Brackbill et al by playing around with the constants in the right hand side. Ended up with constant = <Rho> * Rho_max * 6.3.., this 6.3 looks like a 2*pi.

Looking on the internet, the Laplace formulas was found several times to be Dp = 2*s/R!!! In Brackbill, the 2 is missing.

The benchmark is not validated, but at least the formula and the correct pressure are corrected now.