Obtained the desired pressure by correcting factor in the right hand side.
We obtain this kind of factor: 750*1000*3.35 = <Rho_Mittelwert> * Rho_max * 3.35.

It doesnt match with any formulas found in the literature...it seems that the problem is not only the factors but also that the curvature kappa is too big.

But kappa has been calculated as in Brackbill, Chang, Nagrath, etc... e.g. div.(grad Phi/|grad Phi|).

Needs more investigation to get the right pressure.