IMPORTANT COMMIT => BUG FOUND!
Thanks to Volker, two big stupid mistakes were corrected: pow(..,1.5) instead of 3/2 and the "+" in the post processing formula.

The curvature seems now to be much better. Constant has still to be put, even if it is difficult to see if it plays a role.

Now the pressure obtained is closer 1.12 but still not correct.
Several things could play a role: viscosity, dx, constant. Other problems= L2(u) is increasing instead of decreasing. and u is not 0.

Also, now the post-processing needs to be re-asserted after the correction. Dam Break problem also.

Here are next steps: set back the benchmark as it should be, re-run simulations, check influence of constants, of mu, and try different variants , and check different mesh size. Then move to dam break with this newly corrected CSF and try it, as well as post processing.